# chore(python): Restructure buffer packing to support nulls and improve performance