… python/tests/test_c_array.py
… python/tests/test_iterator.py